+++ internal/provider/resource_entra_group.go
@@ -66,25 +66,15 @@ func (r *NewEntraGroupResource) Schema(ctx context.Context, req resource.SchemaR
 				Validators: []validator.String{
 					stringvalidator.RegexMatches(
 						regexp.MustCompile(`^[\[][æøåÆØÅa-zA-Z0-9 _\-\[\]]+$`),
-						"The name of the Entra group must start with a prefix enclosed in square brackets, and may only contain alphanumeric characters, space ( ), square brackets ([]), underscore (_), and dash (-). The maxiumum length is 64 characters.",
+						"The name of the Entra group must start with a prefix enclosed in square brackets, and may only contain alphanumeric characters, space ( ), square brackets ([]), underscore (_), and dash (-). The maxiumum length is 256 characters.",
 					),
-					stringvalidator.LengthAtMost(64),
+					stringvalidator.LengthAtMost(256),
 				},
 			},
 			"alias": schema.StringAttribute{
-				Required:            true,
-				MarkdownDescription: "The alias of the Entra Group. Must be unique.",
-				Validators: []validator.String{
-					stringvalidator.RegexMatches(
-						regexp.MustCompile(`^[^@()\\\[\]";:<>, ]+$`), // same limitations as mailNickname, see microsoft documentation
-						"The alias is mandatory and it may not contain the characters @ , () \\ [] \" ; : <> SPACE. The max length is 64 characters.",
-					),
-					stringvalidator.LengthAtMost(64),
-				},
-
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
+				Optional:            true,
+				MarkdownDescription: "Alias for the Entra Group. Deprecated and no longer in use.",
+				DeprecationMessage:  "Alias is deprecated and not used in the API. Field will be removed in a future release.",
 			},
 			"description": schema.StringAttribute{
 				Optional:            true,
@@ -143,7 +133,6 @@ func (r *NewEntraGroupResource) Create(ctx context.Context, req resource.CreateR
 
 	entraGroup := tilgangsportalapi.EntraGroup{
 		DisplayName:      data.DisplayName.ValueString(),
-		Alias:            data.Alias.ValueString(),
 		Description:      data.Description.ValueString(),
 		InheritanceLevel: data.InheritanceLevel.ValueString(),
 	}